[!] Control facetting parameters in `ggplot` of `hyperSpec` object

Control facetting parameters in plot made with function ggplot of hyperSpec (ggplot.hyperSpec method) in which parameter wl.range is explicitly specified.

facet_wl(rows = ".", space = "free", scales = "free", ...)

Arguments

rows
Name of factor variable (quoted string) to be used for making rows of facets. The dot "." (default) is used to indicate that there should be no faceting on this dimension.

space
If "fixed", the default, all panels have the same size. If "free_y" their height will be proportional to the length of the y scale; if "free_x" their width will be proportional to the length of the x scale; or if "free" both height and width will vary. This setting has no effect unless the appropriate scales also vary.

scales
Are scales shared across all facets (the default, "fixed"), or do they vary across rows ("free_x"), columns ("free_y"), or both rows and columns ("free")?

...
Further parameters to be passed to facet_grid.

Value

Updated ggplot2 object of class ggplot.

Details

This function is a wrapper for facet_grid and facet_grid(.~.wl.range) can be used instead of facet_wl(".").

See also

Examples

library(spPlot)

p <- ggplot(Spectra2, wl.range = c(min ~ 350, 500~600)) +
     geom_line(aes(color = gr))

#> Error in Re(from): non-numeric argument to function

p

#> Error in eval(expr, envir, enclos): object 'p' not found

p + facet_wl(space = "fixed")

#> Error in eval(expr, envir, enclos): object 'p' not found

p + facet_wl("class")

#> Error in eval(expr, envir, enclos): object 'p' not found

p + facet_wl("gr")

#> Error in eval(expr, envir, enclos): object 'p' not found

# Alternative to `facet_wl()` is:
p + facet_grid(. ~ .wl.range, scales = "free")

#> Error in eval(expr, envir, enclos): object 'p' not found

if (FALSE) {
# If `wl.range` is not specified, ERROR message is displayed:
p2 <- ggplot(Spectra2) + geom_line(aes(color = gr))
p2 + facet_wl()
}