[!] Convert 'plotmath' expresions in labels to text

[!] Convert 'plotmath' expresions in labels to text

label_expr2text(obj, FUN = expr2latex_, ...)

# S3 method for ggplot
label_expr2text(obj, FUN = expr2latex_, ...)

# S3 method for hyperSpec
label_expr2text(obj, FUN = expr2latex_, ...)

# S3 method for plotly_hash
label_expr2text(obj, FUN = expr2latex_, ...)

# S3 method for plotly
label_expr2text(obj, FUN = expr2latex_, ...)

Arguments

obj
Object of approptiate class.

FUN
Function that translates expressions in labels to text or other kind of expressions to be applied, e.g., simsalapar::expr2latex, simsalapar::escapeLatex, expr2latex_.

...
For generic use.

Value

Updated object with expresions in its labels converted to strings.

See also

expr2latex_

Other spPlot utilities: expr2latex_

Other spPlot functions for plotly: plotly_annotation_rm, plotly_modify_hover, plotly_modify_legend, plotly_tidy

Other spPlot functions for spectroscopy and hyperSpec: facet_wl, ggLims, gg_palette, ggplot.hyperSpec

Examples

library(spPlot)
library(hyperSpec)
library(plotly)

# ===== for hyperSpec object ===============================================

labels(chondro)

#> $y
#> expression("y")
#> 
#> $x
#> expression("x")
#> 
#> $clusters
#> expression("clusters")
#> 
#> $spc
#> expression("I / a.u.")
#> 
#> $.wavelength
#> expression(Delta * tilde(nu)/cm^-1)
#> 

labels(label_expr2text(chondro))

#> $y
#> [1] "y"
#> 
#> $x
#> [1] "x"
#> 
#> $clusters
#> [1] "clusters"
#> 
#> $spc
#> [1] "I / a.u."
#> 
#> $.wavelength
#> [1] "Delta * tilde(nu)/cm^-1"
#> 

# NOT RUN {
\donttest{
# without `label_expr2text` labels with expressions are distorted or not plotted at all:

chondro[1:10,,] %>% qplotspc %>% ggplotly
}
# }

# If function `label_expr2text` is applied:
chondro[1:10,,] %>% label_expr2text  %>% qplotspc %>% ggplotly

# ===== for ggplot2 object =================================================

chondro[1:10,,] %>% qplotspc %>% label_expr2text %>% ggplotly

# ===== for plotly object ==================================================

chondro[1:10,,]  %>% qplotspc %>%  ggplotly %>% label_expr2text
# NOT RUN {
\donttest{
## Does not work with newer version of `plotly`:
chondro[1:10,,]  %>% qplotspc %>%  ggplotly %>% plotly_build %>% label_expr2text
}
# }

# ==========================================================================